opal/mca/hwloc: Cleanup dead code.
- This eliminates a lot of stale/dead code that is no longer used
  anywhere in ompi. Most of this (mca's included) have been moved to prrte.

- There are still some left-over api's used by various components (mostly
  btl/smcuda and coll/sm), so hold off removing what they consume.

- Move some functions ompi/mpiext/affinity as it was the only consumer.

- Only filter the cpu-list if the process does its own discovery - IE
  the topology was not provided by pmix/prrte.

(cherry picked from commit e3bfc1eb12de48cc4e854acd95d0aab343faec3d)
