+++ opal/mca/hwloc/base/base.h
@@ -61,12 +61,6 @@ typedef struct {
     int cntr;
 } opal_hwloc_print_buffers_t;
 opal_hwloc_print_buffers_t *opal_hwloc_get_print_buffer(void);
-extern char *opal_hwloc_print_null;
-OPAL_DECLSPEC char *opal_hwloc_base_print_locality(opal_hwloc_locality_t locality);
-
-OPAL_DECLSPEC extern char *opal_hwloc_base_cpu_list;
-OPAL_DECLSPEC extern hwloc_cpuset_t opal_hwloc_base_given_cpus;
-OPAL_DECLSPEC extern char *opal_hwloc_base_topo_file;
 
 /* convenience macro for debugging */
 #define OPAL_HWLOC_SHOW_BINDING(n, v, t)                                                        \
@@ -96,21 +90,6 @@ OPAL_DECLSPEC extern char *opal_hwloc_base_topo_file;
             cache_level = 0;                                   \
         } while (0)
 #endif
-
-OPAL_DECLSPEC opal_hwloc_locality_t opal_hwloc_base_get_relative_locality(hwloc_topology_t topo,
-                                                                          char *cpuset1,
-                                                                          char *cpuset2);
-
-OPAL_DECLSPEC int opal_hwloc_base_set_binding_policy(opal_binding_policy_t *policy, char *spec);
-
-/**
- * Loads opal_hwloc_my_cpuset (global variable in
- * opal/mca/hwloc/hwloc-internal.h) for this process.  opal_hwloc_my_cpuset
- * will be loaded with this process' binding, or, if the process is
- * not bound, use the hwloc root object's (available and online)
- * cpuset.
- */
-OPAL_DECLSPEC void opal_hwloc_base_get_local_cpuset(void);
 
 struct opal_rmaps_numa_node_t {
     opal_list_item_t super;
@@ -153,13 +132,6 @@ OPAL_DECLSPEC extern opal_hwloc_base_mbfa_t opal_hwloc_base_mbfa;
 OPAL_DECLSPEC int opal_hwloc_base_get_topology(void);
 
 /**
- * Set the hwloc topology to that from the given topo file
- */
-OPAL_DECLSPEC int opal_hwloc_base_set_topology(char *topofile);
-
-OPAL_DECLSPEC int opal_hwloc_base_filter_cpus(hwloc_topology_t topo);
-
-/**
  * Free the hwloc topology.
  */
 OPAL_DECLSPEC void opal_hwloc_base_free_topology(hwloc_topology_t topo);
@@ -174,21 +146,6 @@ OPAL_DECLSPEC hwloc_obj_t opal_hwloc_base_get_obj_by_type(hwloc_topology_t topo,
                                                           unsigned int instance,
                                                           opal_hwloc_resource_type_t rtype);
 
-OPAL_DECLSPEC char *opal_hwloc_base_print_binding(opal_binding_policy_t binding);
-
-/**
- * Determine if there is a single cpu in a bitmap.
- */
-OPAL_DECLSPEC bool opal_hwloc_base_single_cpu(hwloc_cpuset_t cpuset);
-
-/**
- * Report a bind failure using the normal mechanisms if a component
- * fails to bind memory -- according to the value of the
- * hwloc_base_bind_failure_action MCA parameter.
- */
-OPAL_DECLSPEC int opal_hwloc_base_report_bind_failure(const char *file, int line, const char *msg,
-                                                      int rc);
-
 /**
  * This function sets the process-wide memory affinity policy
  * according to opal_hwloc_base_map and opal_hwloc_base_mbfa.  It needs
@@ -202,8 +159,6 @@ OPAL_DECLSPEC int opal_hwloc_base_set_process_membind_policy(void);
 OPAL_DECLSPEC int opal_hwloc_base_membind(opal_hwloc_base_memory_segment_t *segs, size_t count,
                                           int node_id);
 
-OPAL_DECLSPEC int opal_hwloc_base_node_name_to_id(char *node_name, int *id);
-
 OPAL_DECLSPEC int opal_hwloc_base_memory_set(opal_hwloc_base_memory_segment_t *segments,
                                              size_t num_segments);
 
@@ -214,32 +169,12 @@ OPAL_DECLSPEC int opal_hwloc_base_memory_set(opal_hwloc_base_memory_segment_t *s
 OPAL_DECLSPEC int opal_hwloc_base_cset2str(char *str, int len, hwloc_topology_t topo,
                                            hwloc_cpuset_t cpuset);
 
-/**
- * Make a prettyprint string for a cset in a map format.
- * Example: [B./..]
- * Key:  [] - signifies socket
- *        / - divider between cores
- *        . - signifies PU a process not bound to
- *        B - signifies PU a process is bound to
- */
-OPAL_DECLSPEC int opal_hwloc_base_cset2mapstr(char *str, int len, hwloc_topology_t topo,
-                                              hwloc_cpuset_t cpuset);
-
-/* get the hwloc object that corresponds to the given processor id  and type */
-OPAL_DECLSPEC hwloc_obj_t opal_hwloc_base_get_pu(hwloc_topology_t topo, int lid,
-                                                 opal_hwloc_resource_type_t rtype);
-
-/* get a string describing the locality of a given process */
-OPAL_DECLSPEC char *opal_hwloc_base_get_locality_string(hwloc_topology_t topo, char *bitmap);
-
 /* extract a location from the locality string */
 OPAL_DECLSPEC char *opal_hwloc_base_get_location(char *locality, hwloc_obj_type_t type,
                                                  unsigned index);
 
 OPAL_DECLSPEC opal_hwloc_locality_t opal_hwloc_compute_relative_locality(char *loc1, char *loc2);
 
-OPAL_DECLSPEC int opal_hwloc_base_topology_set_flags(hwloc_topology_t topology, unsigned long flags,
-                                                     bool io);
 END_C_DECLS
 
 #endif /* OPAL_HWLOC_BASE_H */
