+++ opal/mca/hwloc/base/help-opal-hwloc-base.txt
@@ -19,43 +19,4 @@ message will only be reported at most once per process.
   File:       %s:%d
   Message:    %s
   Severity:   %s
-#
-[invalid binding_policy]
-The specified %s policy is not recognized:
 
-  Policy: %s
-
-Please check for a typo or ensure that the option is a supported
-one.
-#
-[redefining-policy]
-Conflicting directives for binding policy are causing the policy
-to be redefined:
-
-  New policy:   %s
-  Prior policy:  %s
-
-Please check that only one policy is defined.
-#
-[deprecated]
-The following command line option and corresponding MCA parameter have
-been deprecated and replaced as follows:
-
-  Command line option:
-    Deprecated:  %s
-    Replacement: %s
-
-  Equivalent MCA parameter:
-    Deprecated:  %s
-    Replacement: %s
-
-The deprecated forms *will* disappear in a future version of Open MPI.
-Please update to the new syntax.
-#
-[obj-idx-failed]
-Open MPI failed to find a cache of a specified type.  This is a highly
-unusual error; it may indicate a system configuration error.  This
-additional information may be of help:
-
-  Message:     %s
-  Cache level: %d
