+++ opal/mca/hwloc/base/hwloc_base_maffinity.c
@@ -11,61 +11,34 @@
 #include "opal_config.h"
 
 #include "opal/constants.h"
+#include "opal/util/show_help.h"
 
 #include "opal/mca/hwloc/base/base.h"
 #include "opal/mca/hwloc/hwloc-internal.h"
 
-/*
- * Don't use show_help() here (or print any error message at all).
- * Let the upper layer output a relevant message, because doing so may
- * be complicated.
+/**
+ * Report a bind failure using the normal mechanisms if a component
+ * fails to bind memory -- according to the value of the
+ * hwloc_base_bind_failure_action MCA parameter.
  */
-int opal_hwloc_base_set_process_membind_policy(void)
+static int opal_hwloc_base_report_bind_failure(const char *file, int line, const char *msg, int rc)
 {
-    int rc = 0, flags;
-    hwloc_membind_policy_t policy;
-    hwloc_cpuset_t cpuset;
+    static int already_reported = 0;
 
-    /* Make sure opal_hwloc_topology has been set by the time we've
-       been called */
-    if (OPAL_SUCCESS != opal_hwloc_base_get_topology()) {
-        return OPAL_ERR_BAD_PARAM;
+    if (!already_reported && OPAL_HWLOC_BASE_MBFA_SILENT != opal_hwloc_base_mbfa) {
+        char hostname[OPAL_MAXHOSTNAMELEN];
+        gethostname(hostname, sizeof(hostname));
+
+        opal_show_help(
+            "help-opal-hwloc-base.txt", "mbind failure", true, hostname, getpid(), file, line, msg,
+            (OPAL_HWLOC_BASE_MBFA_WARN == opal_hwloc_base_mbfa)
+                ? "Warning -- your job will continue, but possibly with degraded performance"
+                : "ERROR -- your job may abort or behave erraticly");
+        already_reported = 1;
+        return rc;
     }
 
-    /* Set the default memory allocation policy according to MCA
-       param */
-    switch (opal_hwloc_base_map) {
-    case OPAL_HWLOC_BASE_MAP_LOCAL_ONLY:
-        policy = HWLOC_MEMBIND_BIND;
-        flags = HWLOC_MEMBIND_STRICT;
-        break;
-
-    case OPAL_HWLOC_BASE_MAP_NONE:
-    default:
-        policy = HWLOC_MEMBIND_DEFAULT;
-        flags = 0;
-        break;
-    }
-
-    cpuset = hwloc_bitmap_alloc();
-    if (NULL == cpuset) {
-        rc = OPAL_ERR_OUT_OF_RESOURCE;
-    } else {
-        int e;
-        hwloc_get_cpubind(opal_hwloc_topology, cpuset, 0);
-        rc = hwloc_set_membind(opal_hwloc_topology, cpuset, policy, flags);
-        e = errno;
-        hwloc_bitmap_free(cpuset);
-
-        /* See if hwloc was able to do it.  If hwloc failed due to
-           ENOSYS, but the base_map == NONE, then it's not really an
-           error. */
-        if (0 != rc && ENOSYS == e && OPAL_HWLOC_BASE_MAP_NONE == opal_hwloc_base_map) {
-            rc = 0;
-        }
-    }
-
-    return (0 == rc) ? OPAL_SUCCESS : OPAL_ERROR;
+    return OPAL_SUCCESS;
 }
 
 int opal_hwloc_base_memory_set(opal_hwloc_base_memory_segment_t *segments, size_t num_segments)
@@ -109,14 +82,6 @@ out:
     if (OPAL_SUCCESS != rc) {
         return opal_hwloc_base_report_bind_failure(__FILE__, __LINE__, msg, rc);
     }
-    return OPAL_SUCCESS;
-}
-
-int opal_hwloc_base_node_name_to_id(char *node_name, int *id)
-{
-    /* GLB: fix me */
-    *id = atoi(node_name + 3);
-
     return OPAL_SUCCESS;
 }
